requirements.txt
beautifulsoup4
flask
MarkupSafe
openai
psycopg2-binary
python-dotenv
requests
